requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.4.0
python-decouple==3.8
psycopg==3.2.3
dj-database-url==2.3.0
whitenoise==6.8.2
gunicorn==23.0.0
django-filter==24.3
django-redis==5.4.0
python-magic==0.4.27
Pillow==10.4.0
celery==5.3.6
redis==5.0.8
drf-yasg>=1.21.5